OSRC: Samsung G93XFXXU2ERF4 Kernel Update

+++ drivers/base/dma-contiguous.c
@@ -336,7 +336,7 @@ int __init dma_contiguous_reserve_area(phys_addr_t size, phys_addr_t base,
  * global one. Requires architecture specific get_dev_cma_area() helper
  * function.
  */
-struct page *dma_alloc_from_contiguous(struct device *dev, int count,
+struct page *dma_alloc_from_contiguous(struct device *dev, size_t count,
 				       unsigned int align)
 {
 	unsigned long mask, pfn, pageno, start = 0;
@@ -349,10 +349,10 @@ struct page *dma_alloc_from_contiguous(struct device *dev, int count,
 
 	if (align > CONFIG_CMA_ALIGNMENT)
 		align = CONFIG_CMA_ALIGNMENT;
-
+/*
 	pr_debug("%s(cma %p, count %d, align %d)\n", __func__, (void *)cma,
 		 count, align);
-
+*/
 	if (!count)
 		return NULL;
